backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
firebase-admin==6.6.0
mysql-connector-python==9.2.0
python-dotenv==1.0.1
resend==2.5.1
pytest==7.4.3
pytest-cov==4.1.0
